+++ src/twitch/command/handler.ts
@@ -80,7 +80,8 @@ export const onMessageHandlerClosure = (authProvider: RefreshingAuthProvider, ch
                         allowCustomTimeSignature: env.ALLOW_CUSTOM_TIME_SIGNATURE,
                         timeSignatureCC: [env.TIME_SIGNATURE_NUMERATOR_CC, env.TIME_SIGNATURE_DENOMINATOR_CC],
                         repetitionsPerLoop: env.REPETITIONS_PER_LOOP,
-                        allowManualCCMessages: env.ALLOW_MANUAL_CC_MESSAGES
+                        allowManualCCMessages: env.ALLOW_MANUAL_CC_MESSAGES,
+                        allowManualLoops: env.ALLOW_MANUAL_LOOPS
                     },
                     twitch
                 );